misc(generator): move lib.rs to main.rs
